# apidog/dguard-kms-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "D'Guard KMS API",
    "description": "D'Guard KMS 관리자 웹을 위한 인증, 키 관리 및 감사 로그 API입니다.\n\n## 개발 스택\n\n| 영역 | 구성 |\n| --- | --- |\n| Frontend | React 19, TypeScript, MUI, Vite |\n| Backend | Java 21, Spring Boot, Spring Security, Spring Data JPA |\n| Database | PostgreSQL 17, Flyway |\n| 보안 | JWT, PBKDF2-HMAC-SHA256, AES-256-GCM, HMAC-SHA256 |\n| 배포 | Docker Compose, Nginx, GitHub Actions |\n\n## 아키텍처\n\n| 계층 | 역할 |\n| --- | --- |\n| React 관리자 웹 | JWT를 Bearer 헤더에 담아 `/api/*` 요청을 전송 |\n| Nginx | 정적 웹 제공 및 `/api/*`를 Spring Boot로 역방향 프록시 |\n| Spring Boot API | 인증·인가, 키 수명주기, 암복호화, 감사 로그 처리 |\n| PostgreSQL | 사용자 해시, 키 메타데이터·암호화된 키 재료, 감사 로그 저장 |\n\n## 키 상태\n\n| 화면 표시 | API 값 | 의미 |\n| --- | --- | --- |\n| 생성됨 | `CREATED` | 생성되어 활성화 대기 중인 키 |\n| 활성화 | `ACTIVE` | 암복호화와 배포가 가능한 키 |\n| 비활성 | `DEACTIVATED` | 암복호화는 차단하고 회전 정책은 유지하는 키 |\n| 침해 | `COMPROMISED` | 키 유출·침해가 의심되어 폐기만 가능한 키 |\n| 폐기 | `DESTROYED` | 원시 키가 제로화되어 복구할 수 없는 최종 상태 |\n\n배포는 생명주기 상태를 변경하지 않고 별도 운영·감사 이력으로 기록합니다.\n\n## 접근 제어\n\n| 대상 | 권한 |\n| --- | --- |\n| 로그인·로그아웃, Swagger 문서 | 토큰 없이 접근 가능 |\n| 내 정보·세션 연장·키 조회·암복호화 테스트 | 유효한 JWT 필요 |\n| 키 생성·수정·상태 변경·갱신·배포 | `ADMIN` 역할 필요 |\n| 감사 로그 조회·무결성 검증 | `ADMIN` 역할 필요 |\n\n우측 상단 **Authorize**에 `Bearer <JWT>` 형식으로 토큰을 입력하면\n인증이 필요한 API를 Swagger UI에서 호출할 수 있습니다.\n",
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    }
  ],
  "security": [
    {
      "bearerAuth": []
    }
  ],
...
  "paths": {
...
    "/api/notices/{noticeUid}": {
      "get": {
        "tags": [
          "notice-controller"
        ],
        "operationId": "get_1",
        "parameters": [
          {
            "name": "noticeUid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseNoticeResponse"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "notice-controller"
        ],
        "operationId": "update_1",
        "parameters": [
          {
            "name": "noticeUid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "metadata": {
                    "$ref": "#/components/schemas/NoticeUpdateRequest"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": [
                  "metadata"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseNoticeResponse"
                }
              }
            }
          }
        },
        "description": "multipart metadata(JSON)와 files(다중). 기존 첨부 포함 게시글당 10개, 개별 10MiB, 요청 전체 101MiB. 마스터키 AES-256-GCM BYTEA 저장. 검증 실패 시 전체 트랜잭션 롤백."
      },
      "delete": {
        "tags": [
          "notice-controller"
        ],
        "operationId": "delete",
        "parameters": [
          {
            "name": "noticeUid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/notices": {
      "get": {
        "tags": [
          "notice-controller"
        ],
        "operationId": "list",
        "parameters": [
          {
            "name": "title",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "exposeYn",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 20
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponsePageResponseNoticeResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "notice-controller"
        ],
        "operationId": "create_1",
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "metadata": {
                    "$ref": "#/components/schemas/NoticeCreateRequest"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": [
                  "metadata"
                ]
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseNoticeResponse"
                }
              }
            }
          }
        },
        "description": "multipart metadata(JSON)와 files(다중). 기존 첨부 포함 게시글당 10개, 개별 10MiB, 요청 전체 101MiB. 마스터키 AES-256-GCM BYTEA 저장. 검증 실패 시 전체 트랜잭션 롤백."
      }
    },
...
    "/api/files/{fileUid}/download": {
      "get": {
        "tags": [
          "notice-controller"
        ],
        "operationId": "download",
        "parameters": [
          {
            "name": "fileUid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "복호화한 원본 파일",
            "content": {
              "application/octet-stream": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "409": {
            "description": "암호문 변조 또는 지원하지 않는 enc_ver"
          }
        },
        "description": "권한과 GCM 인증 태그를 먼저 검증하고 attachment로 전송. no-store, nosniff. 64KiB씩 출력 후 평문 버퍼 제거. 파일 전체 인증을 위해 최대 10MiB 평문은 일시적으로 메모리에 보관."
      }
    },
    "/api/dashboard/usage-trend": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "키 생성·사용 추이",
        "description": "KST 날짜 기준 일/월별 0 포함 시계열. 생략 시 오늘 포함 최근 30일. 종료일 포함. 키 생성·암호화·복호화 호출 수(실패 포함).",
        "operationId": "usageTrend",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "DAY"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDashboardTrendResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/trends": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "키 생성·사용 추이",
        "description": "KST 날짜 기준 일/월별 0 포함 시계열. 생략 시 오늘 포함 최근 30일. 종료일 포함. 키 생성·암호화·복호화 호출 수(실패 포함).",
        "operationId": "usageTrend_1",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "interval",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "DAY"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseDashboardTrendResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/files/{fileUid}": {
      "delete": {
        "tags": [
          "notice-controller"
        ],
        "operationId": "deleteFile",
        "parameters": [
          {
            "name": "fileUid",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponseVoid"
                }
              }
            }
          }
        }
      }
    },
    "/api/dashboard/expiring": {
      "get": {
        "tags": [
          "Dashboard"
        ],
        "summary": "만료 임박 ACTIVE 키 목록",
        "description": "KST 오늘부터 days일 후까지 포함. expireAt 오름차순, 동일 만료일은 UUID순. 키 값 미반환.",
        "operationId": "expiring",
        "parameters": [
          {
            "name": "days",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 30,
              "minimum": 1,
              "maximum": 365
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/DashboardExpiringKeyResponse"
                      }
                    },
                    "message": {
                      "type": "string"
                    },
                    "errorCode": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "days 범위 오류"
          },
          "401": {
            "description": "미인증"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "NoticeUpdateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "content": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 0
          },
          "exposeYn": {
            "type": "string",
            "pattern": "Y|N"
          }
        },
        "required": [
          "content",
          "title"
        ]
      },
      "ApiResponseNoticeResponse": {
        "type": "object",
        "description": "모든 API의 공통 응답 형식",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청 처리 성공 여부",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/NoticeResponse",
            "description": "API별 응답 데이터. 실패 시 null"
          },
          "message": {
            "type": "string",
            "description": "사용자에게 표시할 처리 결과 메시지",
            "example": "조회에 성공했습니다."
          },
          "errorCode": {
            "type": [
              "string",
              "null"
            ],
            "description": "실패한 경우의 애플리케이션 오류 코드",
            "example": "UNAUTHORIZED"
          }
        }
      },
      "NoticeFileResponse": {
        "type": "object",
        "properties": {
          "fileUid": {
            "type": "string",
            "format": "uuid"
          },
          "originalName": {
            "type": "string"
          },
          "size": {
            "type": "integer",
            "format": "int64"
          },
          "encVer": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "NoticeResponse": {
        "type": "object",
        "properties": {
          "noticeUid": {
            "type": "string",
            "format": "uuid"
          },
          "title": {
            "type": "string"
          },
          "content": {
            "type": "string"
          },
          "exposeYn": {
            "type": "string"
          },
          "viewCount": {
            "type": "integer",
            "format": "int64"
          },
          "createdBy": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "files": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/NoticeFileResponse"
            }
          }
        }
      },
...
      "NoticeCreateRequest": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "content": {
            "type": "string",
            "maxLength": 100000,
            "minLength": 0
          },
          "exposeYn": {
            "type": "string",
            "pattern": "Y|N"
          }
        },
        "required": [
          "content",
          "title"
        ]
      },
...
      "ApiResponseVoid": {
        "type": "object",
        "description": "모든 API의 공통 응답 형식",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청 처리 성공 여부",
            "example": true
          },
          "data": {
            "description": "API별 응답 데이터. 실패 시 null"
          },
          "message": {
            "type": "string",
            "description": "사용자에게 표시할 처리 결과 메시지",
            "example": "조회에 성공했습니다."
          },
          "errorCode": {
            "type": [
              "string",
              "null"
            ],
            "description": "실패한 경우의 애플리케이션 오류 코드",
            "example": "UNAUTHORIZED"
          }
        }
      },
...
      "ApiResponsePageResponseNoticeResponse": {
        "type": "object",
        "description": "모든 API의 공통 응답 형식",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청 처리 성공 여부",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/PageResponseNoticeResponse",
            "description": "API별 응답 데이터. 실패 시 null"
          },
          "message": {
            "type": "string",
            "description": "사용자에게 표시할 처리 결과 메시지",
            "example": "조회에 성공했습니다."
          },
          "errorCode": {
            "type": [
              "string",
              "null"
            ],
            "description": "실패한 경우의 애플리케이션 오류 코드",
            "example": "UNAUTHORIZED"
          }
        }
      },
      "PageResponseNoticeResponse": {
        "type": "object",
        "description": "페이지 단위 조회 결과",
        "properties": {
          "content": {
            "type": "array",
            "description": "현재 페이지의 데이터 목록",
            "items": {
              "$ref": "#/components/schemas/NoticeResponse"
            }
          },
          "page": {
            "type": "integer",
            "format": "int32",
            "description": "0부터 시작하는 현재 페이지 번호",
            "example": 0
          },
          "size": {
            "type": "integer",
            "format": "int32",
            "description": "페이지당 데이터 수",
            "example": 20
          },
          "totalElements": {
            "type": "integer",
            "format": "int64",
            "description": "전체 데이터 수",
            "example": 125
          },
          "totalPages": {
            "type": "integer",
            "format": "int32",
            "description": "전체 페이지 수",
            "example": 7
          }
        }
      },
...
      "ApiResponseDashboardTrendResponse": {
        "type": "object",
        "description": "모든 API의 공통 응답 형식",
        "properties": {
          "success": {
            "type": "boolean",
            "description": "요청 처리 성공 여부",
            "example": true
          },
          "data": {
            "$ref": "#/components/schemas/DashboardTrendResponse",
            "description": "API별 응답 데이터. 실패 시 null"
          },
          "message": {
            "type": "string",
            "description": "사용자에게 표시할 처리 결과 메시지",
            "example": "조회에 성공했습니다."
          },
          "errorCode": {
            "type": [
              "string",
              "null"
            ],
            "description": "실패한 경우의 애플리케이션 오류 코드",
            "example": "UNAUTHORIZED"
          }
        }
      },
      "DashboardTrendPointResponse": {
        "type": "object",
        "description": "차트의 기간별 키 생성·사용 수치",
        "properties": {
          "period": {
            "type": "string",
            "format": "date"
          },
          "keysCreated": {
            "type": "integer",
            "format": "int64"
          },
          "encryptions": {
            "type": "integer",
            "format": "int64"
          },
          "decryptions": {
            "type": "integer",
            "format": "int64"
          },
          "totalOperations": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
      "DashboardTrendResponse": {
        "type": "object",
        "description": "대시보드 라인 차트 시계열",
        "properties": {
          "from": {
            "type": "string",
            "format": "date"
          },
          "to": {
            "type": "string",
            "format": "date"
          },
          "interval": {
            "type": "string"
          },
          "points": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DashboardTrendPointResponse"
            }
          }
        }
      },
...
      "DashboardExpiringKeyResponse": {
        "type": "object",
        "properties": {
          "keyUid": {
            "type": "string",
            "format": "uuid"
          },
          "keyName": {
            "type": "string"
          },
          "algorithm": {
            "type": "string"
          },
          "expireAt": {
            "type": "string",
            "format": "date"
          }
        }
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "description": "로그인 응답의 JWT를 Bearer 토큰으로 입력합니다.",
        "in": "header",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}